+++ otterdog/operations/diff_operation.py
@@ -166,7 +166,7 @@ class DiffOperation(Operation):
             validation_infos,
             validation_warnings,
             validation_errors,
-        ) = await self._validator.validate(expected_org, jsonnet_config.template_dir, self.gh_client)
+        ) = await self._validator.validate(expected_org, jsonnet_config, self.gh_client)
         if validation_errors > 0:
             self.printer.println("Planning aborted due to validation errors.")
             return validation_errors